Update - sweetalerts

+++ blogfront/src/components/PostBox.jsx
@@ -3,7 +3,8 @@ import { api } from "../services/api";
 import { formatDistanceToNow, parseISO } from 'date-fns';
 import { ptBR } from 'date-fns/locale';
 import { FiEdit, FiCheck, FiX, FiTrash2 } from "react-icons/fi";
-import ModalConfirmar from "./ModalConfirmar";
+import ModalConfirmar from "./ModalConfirmar"; 
+import Swal from "sweetalert2";
 
 const formatRelativeTime = (dateString) => {
     if (!dateString) return "Data não disponível";
@@ -19,18 +20,26 @@ const PostBox = () => {
     const [posts, setPosts] = useState([]);
     const [newComment, setNewComment] = useState("");
     const [editingPost, setEditingPost] = useState(null);
-    const [editingComment, setEditingComment] = useState(null);
     const [editedContent, setEditedContent] = useState("");
     const [editedTitle, setEditedTitle] = useState("");
     const [editedCategory, setEditedCategory] = useState("");
-    const [isModalOpen, setIsModalOpen] = useState(false); 
-    const [itemToDelete, setItemToDelete] = useState(null); 
+    const [isModalOpen, setIsModalOpen] = useState(false);  
+    const [itemToDelete, setItemToDelete] = useState(null);  
     const [itemType, setItemType] = useState('');  
+    const [isLoggedIn, setIsLoggedIn] = useState(false);  
+    const [userEmail, setUserEmail] = useState(""); 
+    const [editingComment, setEditingComment] = useState(null); 
+    const [editedCommentContent, setEditedCommentContent] = useState("");
 
     useEffect(() => {
+        const token = localStorage.getItem("Authorization");
+        const email = JSON.parse(localStorage.getItem("user") || '""');
+        setUserEmail(email);
+        setIsLoggedIn(!!token);  
+
         api.get("/posts")
             .then((res) => setPosts(res.data.reverse()))
-            .catch((err) => alert("Deu erro => " + err));
+            .catch((err) => Swal.fire("Erro", "Deu erro => " + err, "error"));
     }, []);
 
     const handleNewComment = (e, postID) => {
@@ -42,7 +51,10 @@ const PostBox = () => {
             window.location.reload();
         })
         .catch((err) => {
-            alert(err.response?.status === 401 ? "Você precisa estar logado para comentar!" : "Erro ao comentar!");
+            Swal.fire({
+                icon: "error",
+                title: err.response?.status === 401 ? "Você precisa estar logado para comentar!" : "Erro ao comentar!",
+            });
         });
         setNewComment("");
     };
@@ -56,7 +68,7 @@ const PostBox = () => {
 
     const savePostEdit = (post) => {
         if (editedTitle === post.title && editedContent === post.content && editedCategory === post.category) {
-            alert("Nenhuma mudança foi feita.");
+            Swal.fire("Aviso", "Nenhuma mudança foi feita.", "info");
             setEditingPost(null);
             return;
         }
@@ -68,34 +80,54 @@ const PostBox = () => {
         }, {
             headers: { Authorization: localStorage.getItem("Authorization") }
         }).then(() => window.location.reload())
-          .catch(() => alert("Erro ao editar post"));
+          .catch(() => Swal.fire("Erro", "Erro ao editar post", "error"));
+    };
+
+    const handleEditComment = (comment) => {
+        setEditingComment(comment.id);
+        setEditedCommentContent(comment.content);
+    };
+
+    const saveCommentEdit = (comment) => {
+        if (editedCommentContent === comment.content) {
+            Swal.fire("Aviso", "Nenhuma mudança foi feita.", "info");
+            setEditingComment(null);
+            return;
+        }
+
+        api.put(`/comments/${comment.id}`, {
+            content: editedCommentContent || comment.content,
+        }, {
+            headers: { Authorization: localStorage.getItem("Authorization") }
+        }).then(() => window.location.reload())
+          .catch(() => Swal.fire("Erro", "Erro ao editar comentário", "error"));
     };
 
     const deletePost = () => {
-        if (itemToDelete) {
+        if (itemToDelete && itemType === 'post') {
             api.delete(`/posts/${itemToDelete.id}`, {
                 headers: { Authorization: localStorage.getItem("Authorization") }
             }).then(() => {
                 setIsModalOpen(false);
                 window.location.reload();
-            }).catch(() => alert("Erro ao excluir post"));
+            }).catch(() => Swal.fire("Erro", "Erro ao excluir post", "error"));
         }
     };
 
     const deleteComment = () => {
-        if (itemToDelete) {
+        if (itemToDelete && itemType === 'comment') {
             api.delete(`/comments/${itemToDelete.id}`, {
                 headers: { Authorization: localStorage.getItem("Authorization") }
             }).then(() => {
                 setIsModalOpen(false);
                 window.location.reload();
-            }).catch(() => alert("Erro ao excluir comentário"));
+            }).catch(() => Swal.fire("Erro", "Erro ao excluir comentário", "error"));
         }
     };
 
     const handleModalClose = () => {
         setIsModalOpen(false);
-        setItemToDelete(null); 
+        setItemToDelete(null);  
         setItemType('');  
     };
 
@@ -109,6 +141,7 @@ const PostBox = () => {
 
     return (
         <div className="space-y-6">
+            {/* Seção de posts */}
             {posts.map((post) => (
                 <div key={post.id} className="bg-gray-50 p-6 rounded-lg shadow-md border border-gray-300 mb-6">
                     <div className="flex justify-between text-sm text-gray-500 mb-4">
@@ -116,19 +149,21 @@ const PostBox = () => {
                             <span className="font-semibold">{post.authorName}</span>
                             <span className="text-gray-500">| Categoria: {post.category}</span>
                         </div>
-                        <div className="flex items-center space-x-2">
-                            <span>{formatRelativeTime(post.dateTime)}</span>
-                            <button onClick={() => handleEditPost(post)} className="text-blue-500 text-sm">
-                                <FiEdit />
-                            </button>
-                            <button onClick={() => {
-                                setItemToDelete(post);
-                                setItemType('post'); 
-                                setIsModalOpen(true);  
-                            }} className="text-red-500 text-sm flex items-center">
-                                <FiTrash2 /> Excluir
-                            </button>
-                        </div>
+                        {isLoggedIn && post.authorEmail === userEmail && (
+                            <div className="flex items-center space-x-2">
+                                <span>{formatRelativeTime(post.dateTime)}</span>
+                                <button onClick={() => handleEditPost(post)} className="text-blue-500 text-sm">
+                                    <FiEdit />
+                                </button>
+                                <button onClick={() => {
+                                    setItemToDelete(post);
+                                    setItemType('post');  
+                                    setIsModalOpen(true); 
+                                }} className="text-red-500 text-sm flex items-center">
+                                    <FiTrash2 /> Excluir
+                                </button>
+                            </div>
+                        )}
                     </div>
 
                     <h1 className="text-2xl font-bold text-[#333333] mb-4">
@@ -163,6 +198,7 @@ const PostBox = () => {
                         <p className="mb-4 text-gray-700">{post.content}</p>
                     )}
 
+                    {/* Seção de comentarios */}
                     <div className="bg-gray-100 p-4 rounded-lg mt-6 border-t-2 border-gray-300">
                         <h2 className="text-lg font-semibold text-[#333333] mb-4">Comentários</h2>
                         <div className="space-y-4">
@@ -171,35 +207,53 @@ const PostBox = () => {
                                     <div key={comment.id} className="border-b border-gray-300 pb-4">
                                         <div className="flex justify-between items-center">
                                             <span className="font-semibold text-gray-800">{comment.author_name}</span>
-                                            <div className="flex items-center space-x-2">
-                                                <span className="text-sm text-gray-500">{formatRelativeTime(comment.datetime)}</span>
-                                                <button onClick={() => {
-                                                    setItemToDelete(comment);
-                                                    setItemType('comment'); 
-                                                    setIsModalOpen(true);  
-                                                }} className="text-red-500 text-sm">
-                                                    <FiTrash2 /> Excluir
-                                                </button>
-                                            </div>
+                                            {isLoggedIn && comment.author_email === userEmail && (
+                                                <div className="flex items-center space-x-2">
+                                                    <button onClick={() => handleEditComment(comment)} className="text-blue-500 text-sm">
+                                                        <FiEdit /> Editar
+                                                    </button>
+                                                    <button onClick={() => {
+                                                        setItemToDelete(comment);
+                                                        setItemType('comment');
+                                                        setIsModalOpen(true);
+                                                    }} className="text-red-500 text-sm">
+                                                        <FiTrash2 /> Excluir
+                                                    </button>
+                                                </div>
+                                            )}
                                         </div>
-                                        <p className="text-gray-600 mt-2">{comment.content}</p>
+                                        {editingComment === comment.id ? (
+                                            <div>
+                                                <textarea value={editedCommentContent} onChange={(e) => setEditedCommentContent(e.target.value)} className="w-full p-2 border rounded-lg"></textarea>
+                                                <button onClick={() => saveCommentEdit(comment)} className="text-green-500 mx-2"><FiCheck /></button>
+                                                <button onClick={() => setEditingComment(null)} className="text-red-500"><FiX /></button>
+                                            </div>
+                                        ) : (
+                                            <p className="text-gray-600 mt-2">{comment.content}</p>
+                                        )}
                                     </div>
                                 ))
                             ) : (
                                 <p className="text-gray-600">Ainda não há comentários.</p>
                             )}
+                            <form onSubmit={(e) => handleNewComment(e, post.id)} className="mt-4 flex">
+                                <input type="text" className="w-full p-2 border border-gray-300 rounded-lg" placeholder="Digite seu comentário..." value={newComment} onChange={(e) => setNewComment(e.target.value)} />
+                                <button type="submit" className="bg-[#007BFF] text-white px-4 py-2 rounded-lg ml-2">Enviar</button>
+                            </form>
                         </div>
                     </div>
                 </div>
             ))}
 
-            {/* Modal de confirmação de exclusão */}
-            <ModalConfirmar 
-                isOpen={isModalOpen}
-                onClose={handleModalClose}
-                onConfirm={handleConfirmDelete}
-                message={itemType === 'post' ? "Você tem certeza que deseja excluir este post?" : "Você tem certeza que deseja excluir este comentário?"}
-            />
+            {/* Modal de confirmação */}
+            {isModalOpen && (
+                <ModalConfirmar
+                    isOpen={isModalOpen}
+                    onClose={handleModalClose}
+                    onConfirm={handleConfirmDelete}
+                    message={itemType === 'post' ? "Tem certeza que deseja excluir este post?" : "Tem certeza que deseja excluir este comentário?"}
+                />
+            )}
         </div>
     );
 };
